+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @vegastack/pages-core
 
+## 0.1.14-next.10
+
+## 0.1.14-next.9
+
 ## 0.1.14-next.8
 
 ## 0.1.14-next.7