+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vegastack/pages-mcp
 
+## 0.1.14-next.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.1.14-next.10
+
+## 0.1.14-next.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.1.14-next.9
+
 ## 0.1.14-next.8
 
 ### Patch Changes