+++ packages/services/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vegastack/pages-services
 
+## 0.1.14-next.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.1.14-next.10
+  - @vegastack/pages-db@0.1.14-next.10
+  - @vegastack/pages-renderer@0.1.14-next.10
+
+## 0.1.14-next.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.1.14-next.9
+  - @vegastack/pages-db@0.1.14-next.9
+  - @vegastack/pages-renderer@0.1.14-next.9
+
 ## 0.1.14-next.8
 
 ### Patch Changes